+++ src/cargo.rs
@@ -1,9 +1,12 @@
 use serde::Deserialize;
+use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::process::{Command, ExitStatus};
 
+use crate::cli::Args;
+use crate::config::Config;
 use crate::errors::*;
-use crate::extensions::CommandExt;
+use crate::extensions::{CommandExt, OutputExt};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum Subcommand {
@@ -52,40 +55,61 @@ impl<'a> From<&'a str> for Subcommand {
     }
 }
 
-#[derive(Debug)]
-pub struct Root {
-    pub path: PathBuf,
+#[derive(Debug, Deserialize)]
+pub struct CargoMetadata {
+    pub workspace_root: PathBuf,
 }
 
-impl Root {
-    pub fn path(&self) -> &Path {
-        &self.path
+impl CargoMetadata {
+    pub fn workspace_root(&self) -> &Path {
+        &self.workspace_root
     }
 }
 
-/// Cargo project root
-pub fn root(cd: Option<&Path>) -> Result<Option<Root>> {
-    #[derive(Deserialize)]
-    struct Manifest {
-        workspace_root: PathBuf,
-    }
+/// Cargo metadata with specific invocation
+pub fn cargo_metadata_with_args(
+    cd: Option<&Path>,
+    args: Option<&Args>,
+) -> Result<Option<CargoMetadata>> {
     let mut command = std::process::Command::new(
         std::env::var("CARGO")
             .ok()
             .unwrap_or_else(|| "cargo".to_string()),
     );
-    command
-        .arg("metadata")
-        .arg("--format-version=1")
-        .arg("--no-deps");
+    command.arg("metadata").arg("--format-version=1");
     if let Some(cd) = cd {
         command.current_dir(cd);
     }
-    let output = command.output()?;
+    if let Some(config) = args {
+        if let Some(ref manifest_path) = config.manifest_path {
+            command.args(["--manifest-path".as_ref(), manifest_path.as_os_str()]);
+        }
+    } else {
+        command.arg("--no-deps");
+    }
+    #[derive(Deserialize)]
+    struct Manifest {
+        workspace_root: PathBuf,
+    }
+    if let Some(cd) = cd {
+        command.current_dir(cd);
+    }
+    let output = command.run_and_get_output(false)?;
+    if !output.status.success() {
+        let mut stderr = std::io::stderr();
+        stderr.write_all(&output.stderr)?;
+        stderr.flush()?;
+        std::process::exit(output.status.code().unwrap_or(1));
+    }
     let manifest: Option<Manifest> = serde_json::from_slice(&output.stdout)?;
-    Ok(manifest.map(|m| Root {
-        path: m.workspace_root,
+    Ok(manifest.map(|m| CargoMetadata {
+        workspace_root: m.workspace_root,
     }))
+}
+
+/// Cargo metadata
+pub fn cargo_metadata(cd: Option<&Path>) -> Result<Option<CargoMetadata>> {
+    cargo_metadata_with_args(cd, None)
 }
 
 /// Pass-through mode